normalize driver __init__ parameters and instantiation
I removed the superfluous 'comm_method' parameter in devices
which supported either LAN or RS232 control.  One method is
the default for each device (we don't particularly care
which one) and the other defaults to None. Also ensured the
parameters are all named identically (ie. 'serial_device'
instead of just 'device').  This is needed to avoid
unnecessary checks in the app, and allow it to be agnostic:
just read the file and instantiate the necessary drivers
with the given parameters.