Implement basic power functionality
For now, the power buttons in the corner will attempt to
turn on/off whatever device is associated with the tab we're
viewing.  Exceptions are caught and result in a status
message which clears after 10 seconds.  At some point I'll
get around to implementing some kind of exception handling
for the input (Command) buttons maybe.

Also at some point, I think I'd like for the corner buttons
to turn on/off ALL display devices in the room.  Each tab
could get its own power buttons maybe toward the bottom
somewhere...

Not sure how best to accomplish but I'd also like it if any
input button pressed would wake the device first if it's in
standby.  Possibly a CommandQueue that's just a series of
commandbutton.Command objects that gets popped one by one
as they are successfully executed?  Would likely need heavy
use of callbacks from drivers to app letting app know when
devices are responding to queries as expected.

Anywho, for now this thing is working better than initially
expected.