add temporary fix for tvs
Temporarily app.power_[on|off]_pressed not check for if
power_on/off() as the samsung tv driver does not return
true/false/anything.  Have started testing with my TV again.
Committing this for now as I have to leave for the day and
tomorrow will get back at this again.